added TCP-connection to control one client from afar (+ a testing program to use it locally, since one can only run one instance of bitcoin locally [well, more or less])
use ./bitcoind makedoublespend and ./bitcoind doublespendlisten for communication:

-on the side of the program to be controlled, call 
./bitcoind doublespendlisten PORT
-on the side of the program to control, call
./bitcoind makedoublespend ADDRESS_1 ADDRESS_2 AMOUNT SERVER PORT
(where address 1, 2 and amount are the bitcoin parameters and server and port the network parameters)
for example to doublespend 11 coins to both of the following addresses, use
 ./bitcoind makedoublespend mkWWAgVjQGu4MtFAdjqaXkrzkywC83JToY mrBXWTkp21NTZQ8XZeS4AR1BPHb5PdVrM8 11 localhost 1234

(where localhost and 1234 are the ports and the machine the other bitcoin program is running on - note that it has to listen for this to work!)

If you want to check locally, use the python program in the directory tcptest. It functions like a buffer in the network, so you can listen on the program, send the data (with makedoublespend) and wait till that's done, then listen on bitcoin and send the data from the program.

To use this, call
python tcptest.py [l|s] PORT
where l means the program will listen, and s means the program will send. To send, it must have before received a transaction first.

So if you want to test locally you would:
$> python tcptest.py l 1234
and then call makedoublespend (as above) with server = localhost and port = 1234
then you call doublespendlisten (as above) with port = 1234
* the port it listens on doesn't have to match the port above, obviously. :)
and then you send from the buffer program:
$> python tcptest.py s 1234